+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 10.3.0
+- **FIX**(screen-resize): Resolve issue causing layers with custom `FractionalTranslation` to be misplaced.
+- **FIX**(paint-editor): Resolve issue where layers didn't resize with the screen.
+- **FIX**(paint-mode): Resolve issue where creating polygons didn't recognize tap events.
+- **FEAT**(paint-editor): Eraser can now also remove existing paintings from other histories.
+
 ## 10.2.8
 - **FIX**(helper-lines): Resolve issue where layers wouldn't release when sharing the same axis.
 - **FIX**(helper-lines): Resolve issue where helperLine configs had no effect.